+++ src/api/local.rs
@@ -41,11 +41,11 @@ impl InnertubeClient {
                 "context": {
                     "client": {
                         "clientName": "ANDROID_VR",
-                        "clientVersion": "1.71.26",
+                        "clientVersion": "1.65.10",
                         "deviceMake": "Oculus",
                         "deviceModel": "Quest 3",
                         "androidSdkVersion": 32,
-                        "userAgent": "com.google.android.apps.youtube.vr.oculus/1.71.26 (Linux; U; Android 12L; eureka-user Build/SQ3A.220605.009.A1) gzip",
+                        "userAgent": "com.google.android.apps.youtube.vr.oculus/1.65.10 (Linux; U; Android 12L; eureka-user Build/SQ3A.220605.009.A1) gzip",
                         "osName": "Android",
                         "osVersion": "12L",
                         "visitorData": "CgtLT21YQTlDUjNqbyjMp-jMBjInCgJCRRIhEh0SGwsMDg8QERITFBUWFxgZGhscHR4fICEiIyQlJiAp",