+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Include channel name while searching in latest videos mode.
 
+###
+- Detect members only status shorts and streams when using Local API.
+
 ## [0.9.0] - 2026-04-28
 ### Added
 - Add Überzug++ as fallback for image rendering.